pyproject.toml
[project]
name = "my_hyperflask_app"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "hyperflask >=0.5.0",
    "hyperflask-users",
    "periodiq", # will be installed from git while an official release is missing
    "pillow",
    "psycopg[binary]",
    "requests"
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-flask>=1.3.0",
    "pytest-asyncio",
    "pytest-playwright>=0.7.1",
    "playwright>=1.55.0",
    "ruff>=0.12.4"
]

[tool.uv.sources]
periodiq = { git = "https://gitlab.com/bersace/periodiq" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "e2e: End-to-end tests using Playwright",
    "unit: Unit tests",
    "integration: Integration tests",
]

# Playwright configuration
[tool.playwright]
browser = "chromium"
headless = true
slow_mo = 0
timeout = 30000
